fix: unwrap hygiene in algo/ + clippy clean across workspace
algo/ unwrap remediation (29 sites, 0 remaining):
- Replace bare .unwrap() with .expect('invariant: ...') in centrality,
  scc, cycle_basis, critical_path, topological_sort, backend helpers
- Fix is_none() + unwrap() in transitive_closure using is_none_or()
- Fix stack.pop().unwrap() in scc.rs and graph_ops.rs

clippy --all-targets -- -D warnings now passes clean (was 43 errors):
- Remove dead code: unused helpers, constants, variants, imports
- Arc -> Rc for non-thread-shared snapshot tests
- Fix println!, needless range loops, unnecessary casts, format-in-println
- Add Incoming/Undirected benchmark cases in adjlist_benchmark
- Fix const asserts, needless borrows, readonly write locks

+++ sqlitegraph-core/examples/compression_analysis.rs
@@ -53,7 +53,7 @@ fn generate_web_graph_pattern(page_count: usize) -> Vec<i64> {
     for hub_id in 1..=hub_count {
         let links_per_hub = 50 + (hub_id as i64 % 51);
         for i in 0..links_per_hub {
-            let target_id = 1i64 + (i as i64 % page_count as i64);
+            let target_id = 1i64 + (i % page_count as i64);
             ids.push(target_id);
         }
     }
